requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
faiss-cpu>=1.7.4
pandas>=2.0.0
python-dotenv>=1.0.0
